go go release

+++ do-libIPN-release.sh
@@ -5,18 +5,21 @@
 
 
 pushd .
-
+mkdir /tmp/IPN
 cd $(mktemp -d /tmp/IPN/libIPN-release.XXXX)
 
 git clone git@gitea.lan:Inventory-Profiles-Next/libIPN.git libIPN
 
-python -m venv ./venv
-. ./venv/bin/activate
-
-pip install pandoc
-pip install pypandoc
-pip install premailer
-pip install pandoc_include
+if [[ ! -e ../venv ]]; then
+  python -m venv ../venv
+  . ../venv/bin/activate
+  pip install pandoc
+  pip install pypandoc
+  pip install premailer
+  pip install pandoc_include
+else
+  . ../venv/bin/activate
+fi
 
 cd libIPN/description
 
@@ -26,4 +29,10 @@ cd ..
 
 ./gradlew --max-workers 32 clean compileKotlin compileJava
 
-#./gradlew --max-workers 4 build publishAllPublicationsToIpnOfficialRepoRepository modrinth curseforge
+GRADLE_ARG="--max-workers 4 build"
+
+if [[ n$IPNEXT_RELEASE != "n" ]]; then
+  GRADLE_ARG="$GRADLE_ARG publishAllPublicationsToIpnOfficialRepoRepository modrinth curseforge"
+fi
+
+./gradlew $GRADLE_ARG
